requirements.txt
pika
pymongo
Pillow
requests
numpy
